+++ lib/ui/explore/ExplorePanel.dart
@@ -17,6 +17,7 @@
 import 'package:flutter/semantics.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:illinois/ext/Explore.dart';
+import 'package:illinois/model/Explore.dart';
 import 'package:illinois/model/Laundry.dart';
 import 'package:illinois/model/Location.dart';
 import 'package:illinois/model/MTD.dart';
@@ -28,7 +29,12 @@ import 'package:illinois/service/Gateway.dart';
 import 'package:illinois/service/Laundries.dart';
 import 'package:illinois/service/MTD.dart';
 import 'package:illinois/service/StudentCourses.dart';
+import 'package:illinois/ui/academics/StudentCourses.dart';
+import 'package:illinois/ui/explore/ExploreBuildingDetailPanel.dart';
+import 'package:illinois/ui/explore/ExploreDiningDetailPanel.dart';
+import 'package:illinois/ui/explore/ExploreEventDetailPanel.dart';
 import 'package:illinois/ui/explore/ExploreSearchPanel.dart';
+import 'package:illinois/ui/laundry/LaundryRoomDetailPanel.dart';
 import 'package:illinois/ui/mtd/MTDStopDeparturesPanel.dart';
 import 'package:illinois/ui/wellness/appointments/AppointmentDetailPanel.dart';
 import 'package:illinois/ui/widgets/FavoriteButton.dart';
@@ -86,6 +92,8 @@ class _ExploreSortKey extends OrdinalSortKey {
 
 class ExplorePanel extends StatefulWidget {
 
+  static const String notifyMapSelect    = "edu.illinois.rokwire.explore.select";
+
   final ExploreItem? initialItem;
   final EventsDisplayType? eventsDisplayType;
   final ExploreFilter? initialFilter;
@@ -120,10 +128,8 @@ class ExplorePanel extends StatefulWidget {
 }
 
 class ExplorePanelState extends State<ExplorePanel>
-    with
-        SingleTickerProviderStateMixin,
-        AutomaticKeepAliveClientMixin<ExplorePanel>
-    implements NotificationsListener {
+  with SingleTickerProviderStateMixin, AutomaticKeepAliveClientMixin<ExplorePanel>
+  implements NotificationsListener {
   
   List<ExploreItem> _exploreItems = [];
   ExploreItem?    _selectedItem;
@@ -171,8 +177,8 @@ class ExplorePanelState extends State<ExplorePanel>
       LocationServices.notifyStatusChanged,
       Localization.notifyStringsUpdated,
       NativeCommunicator.notifyMapSelectExplore,
-      NativeCommunicator.notifyMapSelectLocation,
       NativeCommunicator.notifyMapSelectPOI,
+      NativeCommunicator.notifyMapSelectLocation,
       Auth2UserPrefs.notifyPrivacyLevelChanged,
       Auth2UserPrefs.notifyFavoritesChanged,
       FlexUI.notifyChanged,
@@ -180,6 +186,7 @@ class ExplorePanelState extends State<ExplorePanel>
       StudentCourses.notifyTermsChanged,
       StudentCourses.notifySelectedTermChanged,
       StudentCourses.notifyCachedCoursesChanged,
+      ExplorePanel.notifyMapSelect,
     ]);
 
 
@@ -1256,6 +1263,7 @@ class ExplorePanelState extends State<ExplorePanel>
     String detailsHint = Localization().getStringEx('panel.explore.button.details.hint', '');
     Color? exploreColor;
     Widget? descriptionWidget;
+    bool canDirections = _userLocationEnabled(), canDetail = true;
 
     if (_selectedMapExplore is Explore) {
       title = _selectedMapExplore?.exploreTitle;
@@ -1266,6 +1274,7 @@ class ExplorePanelState extends State<ExplorePanel>
         detailsHint = Localization().getStringEx('panel.explore.button.bus_schedule.hint', '');
         descriptionWidget = _buildStopDescription();
       }
+      canDetail = !(_selectedMapExplore is ExplorePOI);
     }
     else if  (_selectedMapExplore is List<Explore>) {
       String? exploreName = ExploreExt.getExploresListDisplayTitle(_selectedMapExplore);
@@ -1276,9 +1285,11 @@ class ExplorePanelState extends State<ExplorePanel>
     }
     else {
       exploreColor = Styles().colors?.white;
+      canDirections = canDetail = false;
     }
 
     double buttonWidth = (MediaQuery.of(context).size.width - (40 + 12)) / 2;
+
     return Stack(clipBehavior: Clip.hardEdge, children: <Widget>[
       (_mapAllowed == true) ? MapWidget(
         onMapCreated: _onNativeMapCreated,
@@ -1290,31 +1301,29 @@ class ExplorePanelState extends State<ExplorePanel>
             Padding(padding: EdgeInsets.symmetric(horizontal: 20, vertical: 10), child:
               Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
                 Padding(padding: EdgeInsets.only(right: 10), child:
-                  Text(title ?? '', overflow: TextOverflow.ellipsis, style: TextStyle(fontFamily: Styles().fontFamilies!.extraBold, fontSize: 20, color: Styles().colors!.fillColorPrimary, )),
+                  Text(title ?? '', maxLines: 1, overflow: TextOverflow.ellipsis, style: TextStyle(fontFamily: Styles().fontFamilies!.extraBold, fontSize: 20, color: Styles().colors!.fillColorPrimary, )),
                 ),
                 (descriptionWidget != null) ?
                   Row(children: <Widget>[
-                    Text(description ?? "", overflow: TextOverflow.ellipsis, style: TextStyle(fontFamily: Styles().fontFamilies!.medium, fontSize: 16, color: Colors.black38,)),
+                    Text(description ?? "", maxLines: 1, overflow: TextOverflow.ellipsis, style: TextStyle(fontFamily: Styles().fontFamilies!.medium, fontSize: 16, color: Colors.black38,)),
                     descriptionWidget
                   ]) :
                   Text(description ?? "", overflow: TextOverflow.ellipsis, style: TextStyle(fontFamily: Styles().fontFamilies!.medium, fontSize: 16, color: Colors.black38,)),
                 Container(height: 8,),
                 Row(children: <Widget>[
-                  _userLocationEnabled() ? Row(children: <Widget>[
-                    SizedBox(width: buttonWidth, child:
-                      RoundedButton(
-                        label: Localization().getStringEx('panel.explore.button.directions.title', 'Directions'),
-                        hint: Localization().getStringEx('panel.explore.button.directions.hint', ''),
-                        backgroundColor: Colors.white,
-                        padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-                        fontSize: 16.0,
-                        textColor: Styles().colors!.fillColorPrimary,
-                        borderColor: Styles().colors!.fillColorSecondary,
-                        onTap: _onTapMapExploreDirections
-                      ),
+                  SizedBox(width: buttonWidth, child:
+                    RoundedButton(
+                      label: Localization().getStringEx('panel.explore.button.directions.title', 'Directions'),
+                      hint: Localization().getStringEx('panel.explore.button.directions.hint', ''),
+                      backgroundColor: Colors.white,
+                      padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                      fontSize: 16.0,
+                      textColor: canDirections ? Styles().colors!.fillColorPrimary : Styles().colors!.surfaceAccent,
+                      borderColor: canDirections ? Styles().colors!.fillColorSecondary : Styles().colors!.surfaceAccent,
+                      onTap: _onTapMapExploreDirections
                     ),
-                    Container(width: 12,),
-                  ]) : Container(),
+                  ),
+                  Container(width: 12,),
                   SizedBox(width: buttonWidth, child:
                     RoundedButton(
                       label: detailsLabel,
@@ -1322,8 +1331,8 @@ class ExplorePanelState extends State<ExplorePanel>
                       backgroundColor: Colors.white,
                       padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8),
                       fontSize: 16.0,
-                      textColor: Styles().colors!.fillColorPrimary,
-                      borderColor: Styles().colors!.fillColorSecondary,
+                      textColor: canDetail ? Styles().colors!.fillColorPrimary : Styles().colors!.surfaceAccent,
+                      borderColor: canDetail ? Styles().colors!.fillColorSecondary : Styles().colors!.surfaceAccent,
                       onTap: _onTapMapExploreDetail,
                     ),
                   ),
@@ -1379,61 +1388,98 @@ class ExplorePanelState extends State<ExplorePanel>
 
   void _selectMapExplore(dynamic explore) {
     if (explore != null) {
-      _refresh(() { _selectedMapExplore = explore;});
+      _nativeMapController?.markPOI(((explore is ExplorePOI) && StringUtils.isEmpty(explore.placeId)) ? explore : null);
+      _refresh(() { _selectedMapExplore = explore; });
       _updateSelectedMapStopRoutes();
       _mapExploreBarAnimationController.forward();
     }
     else if (_selectedMapExplore != null) {
+      _nativeMapController?.markPOI(null);
       _mapExploreBarAnimationController.reverse().then((_){
-        _refresh(() { _selectedMapExplore = null;});
+        _refresh(() { _selectedMapExplore = null; });
         _updateSelectedMapStopRoutes();
       });
     }
   }
 
   void _onTapMapExploreDirections() {
-      Analytics().logSelect(target: 'Directions');
+    Analytics().logSelect(target: 'Directions');
+    if (_userLocationEnabled()) {
       dynamic explore = _selectedMapExplore;
-      _mapExploreBarAnimationController.reverse().then((_){
-        _refresh(() { _selectedMapExplore = null;});
-      });
+      _selectMapExplore(null);
       if (explore != null) {
-        NativeCommunicator().launchExploreMapDirections(target: explore);
+        String? travelMode;
+        if (explore is List) {
+          dynamic exploreEntry = (0 < explore.length) ? explore.first : null;
+          travelMode = ((exploreEntry is MTDStop) || (exploreEntry is ExplorePOI)) ? 'transit' : null;
+        }
+        else {
+          travelMode = ((explore is MTDStop) || (explore is ExplorePOI)) ? 'transit' : null;
+        }
+        NativeCommunicator().launchExploreMapDirections(target: explore, options: (travelMode != null) ? {
+          'travelMode': travelMode
+        } : null);
       }
+    }
+    else {
+      AppAlert.showMessage(context, Localization().getStringEx("panel.explore.directions.na.msg", "You need to enable location services in order to get navigation directions."));
+    }
   }
   
   void _onTapMapExploreDetail() {
-      Analytics().logSelect(target: (_selectedMapExplore is MTDStop) ? 'Bus Schedule' : 'Details');
+    Analytics().logSelect(target: (_selectedMapExplore is MTDStop) ? 'Bus Schedule' : 'Details');
 
-      dynamic explore = _selectedMapExplore;
-      _mapExploreBarAnimationController.reverse().then((_){
-        _refresh(() { _selectedMapExplore = null;});
-      });
+    Route? route;
+    String? message;
+    if (_selectedMapExplore is Event) {
+      if (_selectedMapExplore.isGameEvent) {
+        route = CupertinoPageRoute(builder: (context) => AthleticsGameDetailPanel(gameId: _selectedMapExplore.speaker, sportName: _selectedMapExplore.registrationLabel,),);
+      }
+      else if (_selectedMapExplore.isComposite) {
+        route = CupertinoPageRoute(builder: (context) => CompositeEventsDetailPanel(parentEvent: _selectedMapExplore),);
+      }
+      else {
+        route = CupertinoPageRoute(builder: (context) => ExploreEventDetailPanel(event: _selectedMapExplore, initialLocationData: _locationData),);
+      }
+    }
+    else if (_selectedMapExplore is Dining) {
+      route = CupertinoPageRoute(builder: (context) => ExploreDiningDetailPanel(dining: _selectedMapExplore, initialLocationData: _locationData),);
+    }
+    else if (_selectedMapExplore is LaundryRoom) {
+      route = CupertinoPageRoute(builder: (context) => LaundryRoomDetailPanel(room: _selectedMapExplore),);
+    }
+    else if (_selectedMapExplore is Game) {
+      route = CupertinoPageRoute(builder: (context) => AthleticsGameDetailPanel(game: _selectedMapExplore),);
+    }
+    else if (_selectedMapExplore is Building) {
+      route = CupertinoPageRoute(builder: (context) => ExploreBuildingDetailPanel(building: _selectedMapExplore),);
+    }
+    else if (_selectedMapExplore is MTDStop) {
+      route = CupertinoPageRoute(builder: (context) => MTDStopDeparturesPanel(stop: _selectedMapExplore,),);
+    }
+    else if (_selectedMapExplore is StudentCourse) {
+      route = CupertinoPageRoute(builder: (context) => StudentCourseDetailPanel(course: _selectedMapExplore,),);
+    }
+    else if (_selectedMapExplore is Appointment) {
+      route = CupertinoPageRoute(builder: (context) => AppointmentDetailPanel(appointment: _selectedMapExplore),);
+    }
+    else if (_selectedMapExplore is ExplorePOI) {
+      message = Localization().getStringEx("panel.explore.details.na.msg", "Details are not available for custom points of interests.");
+    }
+    else if (_selectedMapExplore is Explore) {
+      route = CupertinoPageRoute(builder: (context) => ExploreDetailPanel(explore: _selectedMapExplore, initialLocationData: _locationData,),);
+    }
+    else if (_selectedMapExplore is List<Explore>) {
+      route = CupertinoPageRoute(builder: (context) => ExploreListPanel(explores: _selectedMapExplore),);
+    }
 
-      if (explore is Explore) {
-        Event? event = (explore is Event) ? explore : null;
-        if (event?.isGameEvent ?? false) {
-          Navigator.push(context, CupertinoPageRoute(builder: (context) =>
-              AthleticsGameDetailPanel(gameId: event!.speaker, sportName: event.registrationLabel,)));
-        }
-        else if(explore is Game) {
-          Navigator.push(context, CupertinoPageRoute(builder: (context) =>
-              AthleticsGameDetailPanel(game: explore)));
-        }
-        else if (explore is MTDStop) {
-          Navigator.push(context, CupertinoPageRoute(builder: (context) => MTDStopDeparturesPanel(stop: explore,)));
-        }
-        else if(explore is Appointment) {
-          Navigator.push(context, CupertinoPageRoute(builder: (context) => AppointmentDetailPanel(appointment: explore)));
-        }
-        else {
-          Navigator.push(context, CupertinoPageRoute(builder: (context) =>
-            ExploreDetailPanel(explore: explore, initialLocationData: _locationData,)));
-        }
-      }
-      else if (explore is List<Explore>) {
-        Navigator.push(context, CupertinoPageRoute(builder: (context) => ExploreListPanel(explores: explore)));
-      }
+    if (route != null) {
+      _selectMapExplore(null);
+      Navigator.push(context, route);
+    }
+    else if (message != null) {
+      AppAlert.showMessage(context, message);
+    }
   }
 
   void _updateSelectedMapStopRoutes() {
@@ -1830,7 +1876,7 @@ class ExplorePanelState extends State<ExplorePanel>
   void _viewStateFarmPoi() {
     Analytics().logSelect(target: "State Farm Wayfinding");
     if (_nativeMapController != null) {
-      _nativeMapController!.viewPoi({
+      _nativeMapController!.viewPOI({
         'latitude': Config().stateFarmWayfinding['latitude'],
         'longitude': Config().stateFarmWayfinding['longitude'],
         'zoom': Config().stateFarmWayfinding['zoom'],
@@ -1879,11 +1925,11 @@ class ExplorePanelState extends State<ExplorePanel>
     else if (name == NativeCommunicator.notifyMapSelectExplore) {
       _onNativeMapSelectExplore(param);
     }
-    else if (name == NativeCommunicator.notifyMapSelectLocation) {
-      _onNativeMapSelectLocation(param);
-    }
     else if (name == NativeCommunicator.notifyMapSelectPOI) {
       _onNativeMapSelectPOI(param);
+    }
+    else if (name == NativeCommunicator.notifyMapSelectLocation) {
+      _onNativeMapSelectLocation(param);
     }
     else if(name == Storage.offsetDateKey){
       _loadExplores();
@@ -1921,6 +1967,15 @@ class ExplorePanelState extends State<ExplorePanel>
         _loadExplores();
       }
     }
+    else if (name == ExplorePanel.notifyMapSelect) {
+      if (mounted) {
+        setState(() {
+          _displayType = ListMapDisplayType.Map;
+          _mapAllowed = true;
+          _selectedItem = param;
+        });
+      }
+    }
   }
 
   void _updateLocationServicesStatus() {
@@ -1944,7 +1999,7 @@ class ExplorePanelState extends State<ExplorePanel>
 
   void _onNativeMapSelectExplore(Map<String, dynamic>? params) {
     int? mapId = (params != null) ? JsonUtils.intValue(params['mapId']) : null;
-    if (_nativeMapController!.mapId == mapId) {
+    if (_nativeMapController?.mapId == mapId) {
       dynamic explore;
       dynamic exploreJson = (params != null) ? params['explore'] : null;
       if (exploreJson is Map) {
@@ -1960,23 +2015,9 @@ class ExplorePanelState extends State<ExplorePanel>
     }
   }
   
-  void _onNativeMapSelectLocation(Map<String, dynamic>? params) {
-    int? mapId = (params != null) ? JsonUtils.intValue(params['mapId']) : null;
-    if (_nativeMapController!.mapId == mapId) {
-      LatLng? location = (params != null) ? LatLng.fromJson(JsonUtils.mapValue(params['location'])) : null;
-      MTDStop? mtdStop = (location != null) ? MTD().stops?.findStop(location: location, locationThresholdDistance: 25 /*in meters*/) : null;
-      if (mtdStop != null) {
-        _selectMapExplore(mtdStop);
-      }
-      else {
-        _selectMapExplore(null);
-      }
-    }
-  }
-
   void _onNativeMapSelectPOI(Map<String, dynamic>? params) {
     int? mapId = (params != null) ? JsonUtils.intValue(params['mapId']) : null;
-    if (_nativeMapController!.mapId == mapId) {
+    if (_nativeMapController?.mapId == mapId) {
       Map<String, dynamic>? poi = (params != null) ? JsonUtils.mapValue(params['poi']) : null;
       String? poiName = (poi != null) ? JsonUtils.stringValue(poi['name']) : null;
       LatLng? poiLocation = (poi != null) ? LatLng.fromJson(JsonUtils.mapValue(poi['location'])) : null;
@@ -1986,12 +2027,28 @@ class ExplorePanelState extends State<ExplorePanel>
           mtdStop = MTD().stops?.findStop(name: poiName) ??
             MTD().stops?.findStop(location: poiLocation, locationThresholdDistance: 10 /*in meters*/);
         }
-        if (mtdStop != null) {
-          _selectMapExplore(mtdStop);
-        }
+        _selectMapExplore(mtdStop ?? ExplorePOI.fromJson(poi));
       }
     }
   }
+
+  void _onNativeMapSelectLocation(Map<String, dynamic>? params) {
+    int? mapId = (params != null) ? JsonUtils.intValue(params['mapId']) : null;
+    if (_nativeMapController?.mapId == mapId) {
+      LatLng? location = (params != null) ? LatLng.fromJson(JsonUtils.mapValue(params['location'])) : null;
+      MTDStop? mtdStop = (location != null) ? MTD().stops?.findStop(location: location, locationThresholdDistance: 25 /*in meters*/) : null;
+      if (mtdStop != null) {
+        _selectMapExplore(mtdStop);
+      }
+      else if (_selectedMapExplore != null) {
+        _selectMapExplore(null);
+      }
+      else if (location?.isValid ?? false){
+        _selectMapExplore(ExplorePOI(location: ExploreLocation(latitude: location?.latitude, longitude: location?.longitude)));
+      }
+    }
+  }
+
 }
 
 ////////////////////